refactor: decompose ScheMatiQRunner into focused pipeline modules
Break down the 2,528-line god class into 7 focused modules under
pipeline/ (llm_factory, callbacks, config_handler, schema_discovery,
value_extraction, data_query, __init__). ScheMatiQRunner stays as
the public facade with identical API — no caller changes needed.

Also: deduplicate enforce_release_llm_config and is_local_path across
services, remove dead _find_paper_path in reextraction_service, and
fix O(n) progress monitoring in upload_document_processor (seek-based).
