+++ backend/app/services/pipeline/__init__.py
@@ -1,0 +1,26 @@
+"""Pipeline modules for ScheMatiQ execution."""
+
+from .llm_factory import build_llm_interface, enforce_release_llm_config
+from .callbacks import create_value_extracted_callback, create_warning_callback, start_heartbeat
+from .config_handler import validate_config, convert_config_to_schematiq_format, resolve_docs_paths
+from .schema_discovery import run_schema_discovery
+from .value_extraction import run_value_extraction, process_suggested_values
+from .data_query import compute_statistics, get_status, get_schema, get_data
+
+__all__ = [
+    "build_llm_interface",
+    "enforce_release_llm_config",
+    "create_value_extracted_callback",
+    "create_warning_callback",
+    "start_heartbeat",
+    "validate_config",
+    "convert_config_to_schematiq_format",
+    "resolve_docs_paths",
+    "run_schema_discovery",
+    "run_value_extraction",
+    "process_suggested_values",
+    "compute_statistics",
+    "get_status",
+    "get_schema",
+    "get_data",
+]
